% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.stance_result}
\alias{as.data.frame.stance_result}
\title{Convert stance_result to data frame}
\usage{
\method{as.data.frame}{stance_result}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{stance_result} object}

\item{row.names}{Optional NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{Logical. If TRUE, setting row names and converting column names to syntactic names is optional.}

\item{...}{Additional arguments passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}.}
}
\value{
A Data frame
}
\description{
Extracts the summary table from a \code{stance_result} object and returns it as
a data frame. This allows for easy further analysis using standard R tools.
}
