% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{inspect}
\alias{inspect}
\title{Inspect stance_result contents}
\usage{
inspect(x, ...)
}
\arguments{
\item{x}{A \code{stance_result} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The input object (invisibly).
}
\description{
Inspect stance_result contents
}
