% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stancer_available_languages.R
\name{stancer_available_languages}
\alias{stancer_available_languages}
\title{Get available languages for stancer}
\usage{
stancer_available_languages()
}
\value{
Character vector of supported language codes
}
\description{
Get available languages for stancer
}
\examples{
stancer_available_languages()
}
