% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{programming_tweets}
\alias{programming_tweets}
\title{Programming Tweets Dataset}
\format{
A tibble with 35 rows and 1 column:
\describe{
\item{tweet}{Text of the tweet}
}
}
\usage{
programming_tweets
}
\description{
A collection of 35 fictional tweets about Julia, Python, and R
for testing stance analysis.
}
\keyword{datasets}
